# simd: add SIMD SVE implementation of L2 distance function