… adapters/repos/db/vector/hnsw/distancer/asm/l2_neon_arm64.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_neon_arm64.s
… adapters/repos/db/vector/hnsw/distancer/asm/l2_stub_arm64.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_sve_arm64.go
… adapters/repos/db/vector/hnsw/distancer/asm/l2_sve_arm64.s