### What's being changed:

Difference between NEON and SVE
```
goos: linux
goarch: arm64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
                                      │ _2_neon.txt  │             _3_sve.txt             │
                                      │    sec/op    │   sec/op     vs base               │
_L2_PureGo_VS_SIMD/vector_dim=1-16       3.015n ± 0%   3.003n ± 0%   -0.38% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=2-16       3.054n ± 0%   2.983n ± 0%   -2.32% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=3-16       3.293n ± 0%   3.468n ± 0%   +5.31% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=4-16       4.672n ± 0%   4.701n ± 0%   +0.62% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=5-16       4.074n ± 2%   4.239n ± 0%   +4.06% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=16-16      7.465n ± 0%   8.636n ± 0%  +15.69% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=31-16      11.51n ± 0%   13.34n ± 0%  +15.90% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=32-16      9.267n ± 0%   9.147n ± 0%   -1.29% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=35-16      12.24n ± 1%   11.76n ± 0%   -3.88% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=64-16      13.33n ± 0%   11.18n ± 0%  -16.16% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=67-16      15.62n ± 1%   14.61n ± 0%   -6.47% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=128-16     21.82n ± 0%   16.45n ± 0%  -24.61% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=130-16     23.41n ± 0%   18.39n ± 0%  -21.44% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=256-16     34.93n ± 0%   27.21n ± 0%  -22.12% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=260-16     36.39n ± 0%   33.56n ± 0%   -7.76% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=384-16     48.03n ± 0%   40.12n ± 0%  -16.46% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=390-16     51.03n ± 0%   45.07n ± 0%  -11.68% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=768-16     86.97n ± 0%   64.41n ± 0%  -25.94% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=777-16     90.08n ± 0%   67.13n ± 0%  -25.48% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=1000-16   111.90n ± 0%   80.58n ± 0%  -27.99% (p=0.002 n=6)
_L2_PureGo_VS_SIMD/vector_dim=1536-16    165.0n ± 0%   118.3n ± 0%  -28.33% (p=0.002 n=6)
geomean                                  18.14n        16.22n       -10.57%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.